feat: 솔로러닝에 페이스 추가 및 솔로코스러닝에 기록저장 및 진행률 추가
