+++ component/Modal/PrepareRunModal/PrepareSoloCourse.tsx
@@ -1,175 +1,95 @@
-import {
-  Text,
-  StyleSheet,
-  View,
-  TouchableOpacity,
-  TextInput,
-} from "react-native";
+import { Text, StyleSheet, View } from "react-native";
+import Button from "../../Button";
 import { useRunningStore } from "@/store/useRunningStore";
 import { useShallow } from "zustand/react/shallow";
 import { useCourseStore } from "@/store/useCourseStore";
-import Button from "@/component/Button";
-import { useEffect, useState } from "react";
-import Input from "@/component/Input";
-import { useStomp } from "@/hooks/running/useStomp";
-import useInterval from "@/hooks/running/useInterval";
 import { useLocationStore } from "@/store/useLocationStore";
+import { useStomp } from "@/hooks/running/useStomp";
+import MyLocation from "@/assets/images/svg/Mylocation";
+import { useEffect } from "react";
+import { useStompStore } from "@/store/useStompStore";
 const PrepareSoloCourse = () => {
-  const { setRunningStatus, targetPace, setTargetPace } = useRunningStore(
+  const { selectedCourseId, setSelectedCourseId } = useCourseStore(
     useShallow((state) => ({
-      targetPace: state.targetPace,
-      setRunningStatus: state.setRunningStatus,
-      setTargetPace: state.setTargetPace,
-
+      selectedCourseId: state.selectedCourseId,
+      setSelectedCourseId: state.setSelectedCourseId,
     }))
   );
-  const myLocation = useLocationStore(state => state.myLocation);
-    const stompLocation = useLocationStore(state => state.stompLocation);
-  const [avaragePace, setAveragePace] = useState("5'30\"");
-  const [minute, setMinute] = useState("0");
-  const [second, setSecond] = useState("0");
-  const { sendLocation } = useStomp();
-
-  const handlePreset = (preset: string) => {
-    setTargetPace(preset);
-    const [min, sec] = preset.split(/['"]/).map(String);
-    setMinute(min);
-    setSecond(sec);
-  };
- useInterval(
-    () => {
-      sendLocation(myLocation);
-    },
-    myLocation ? 1000 : null
+  const { sendLocation, connected } = useStomp();
+  const myLocation = useLocationStore((state) => state.myLocation);
+  const client = useStompStore((state) => state.client);
+  const { setRunningStatus, setRunningInfo } = useRunningStore(
+    useShallow((state) => ({
+      setRunningStatus: state.setRunningStatus,
+      setRunningInfo: state.setRunningInfo,
+    }))
   );
-  const handleSecondChange = (text: string) => {
-  // 숫자만 필터링
-  const onlyNumbers = text.replace(/[^0-9]/g, '');
-
-  if (onlyNumbers.length > 2) return;
-
-  setSecond(onlyNumbers);
-
-  if (onlyNumbers.length === 2) {
-    const num = parseInt(onlyNumbers, 10);
-    if (num > 59) {
-      setSecond('59'); // 자동으로 59로 고정
-    }
-  }
-};
-
   useEffect(() => {
-    setTargetPace(`${minute}'${second}"`);
-  }, [minute, second]);
+    if (client && myLocation && connected) {
+      sendLocation(myLocation);
+    }
+  }, [myLocation, connected]);
+  const stompLocation = useLocationStore((state) => state.stompLocation);
   return (
     <>
-      <Text style={styles.title}>목표 페이스 설정</Text>
-      <Text style={styles.subtitle}>경쟁자 평균 페이스: {avaragePace}/km</Text>
-
-      <View style={styles.buttonRow}>
-        {["5'00\"", "5'30\"", "6'00\""].map((preset, idx) => (
-          <Button
-            style={[styles.presetButton]}
-            theme={`${targetPace === preset ? "default" : "secondary"}`}
-            onPress={() => handlePreset(preset)}
-            key={idx}
-          >
-            {preset}
-          </Button>
-        ))}
+      <Text style={styles.modalDepscription2}>
+        현재 위치를 기준으로 러닝을 시작합니다.
+      </Text>
+      <Text style={styles.modalDepscription}>러닝을 시작하시겠어요?</Text>
+    
+      {stompLocation?.runningStatus === "ONGOING" ? (
+      <View style={styles.buttonContainer}>
+        <Button
+          onPress={() => {
+            setRunningStatus("countdown");
+          }}
+          style={styles.button}
+        >
+          러닝 시작!
+        </Button>
+        <Button
+          onPress={() => {
+            setRunningStatus("idle");
+            if (selectedCourseId === "solo") {
+              setSelectedCourseId("");
+              client?.deactivate();
+            }
+          }}
+          style={styles.button}
+        >
+          취소
+        </Button>
       </View>
-
-      <View style={{ flexDirection: "row", alignItems: "flex-start" }}>
-        <TextInput
-          style={[styles.input, { width: 50, textAlign: "center" }]}
-          value={minute}
-          onChangeText={setMinute}
-          keyboardType="numeric"
-          maxLength={2}
-          placeholder="분"
-        />
-        <Text style={{ marginHorizontal: 4 }}>'</Text>
-        <TextInput
-          style={[styles.input, { width: 50, textAlign: "center" }]}
-          value={second}
-          onChangeText={handleSecondChange}
-          keyboardType="numeric"
-          maxLength={2}
-          placeholder="초"
-        />
-        <Text style={{ marginLeft: 4 }}>''</Text>
-      </View>
-        {stompLocation?.runningStatus === 'ONGOING' ?
-      (<Button
-        style={styles.runButton}
-        onPress={() => {
-          setRunningStatus("countdown");
-        }}
-      >
-        러닝 시작하기
-      </Button>): (<Text style = {{color:'red', marginBottom:4, fontSize: 18}}>코스에서 벗어났습니다</Text>)}
+      ): (
+        <Text style={{ color: "red", marginBottom: 4, fontSize: 18 }}>
+          코스에서 벗어났습니다
+        </Text>
+      )}
     </>
   );
 };
 export default PrepareSoloCourse;
 const styles = StyleSheet.create({
-  overlay: {
-    flex: 1,
-    backgroundColor: "rgba(0,0,0,0.3)",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  modal: {
-    width: "85%",
-    backgroundColor: "#fff",
-    borderRadius: 16,
-    padding: 24,
-    alignItems: "center",
-  },
-  title: {
+  modalDepscription: {
+    fontWeight: 500,
     fontSize: 20,
-    fontWeight: "700",
-    marginBottom: 8,
+    lineHeight: 28,
+    color: "#000000",
   },
-  subtitle: {
-    fontSize: 14,
-    color: "#555",
-    marginBottom: 20,
+  modalDepscription2: {
+    fontWeight: 300,
+    fontSize: 16,
+    lineHeight: 28,
+    color: "#000000",
   },
-  buttonRow: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    marginBottom: 16,
+  buttonContainer: {
     width: "100%",
+    gap: 12,
+    paddingHorizontal: 48,
+    marginTop: 24,
   },
-  presetButton: {
+  button: {
     flex: 1,
-    paddingVertical: 10,
-    marginHorizontal: 4,
-    borderRadius: 8,
-    alignItems: "center",
-  },
-  presetButtonActive: {
-    backgroundColor: "#6200EE",
-  },
-  presetButtonText: {
-    color: "#333",
-    fontWeight: "600",
-  },
-  presetButtonTextActive: {
-    color: "#fff",
-  },
-  input: {
-    width: "100%",
-    borderWidth: 1,
-    borderColor: "#ddd",
-    borderRadius: 12,
-    paddingVertical: 10,
-    paddingHorizontal: 12,
-    fontSize: 18,
-    fontWeight: "bold",
-    marginBottom: 24,
-    textAlign: "center",
   },
   runButton: {
     width: "100%",
@@ -177,10 +97,5 @@ const styles = StyleSheet.create({
     paddingVertical: 14,
     borderRadius: 12,
     alignItems: "center",
-  },
-  runButtonText: {
-    color: "#fff",
-    fontSize: 16,
-    fontWeight: "bold",
   },
 });
