+++ component/Modal/RunningModal/SoloRunning.tsx
@@ -1,8 +1,10 @@
+import { SoloRunningText } from "@/component/SoloRunningText";
 import RunningInfo from "./RunningInfo";
 import { useSoloRunning } from "@/hooks/running/useSoloRunning";
 import { convertSpeedToPace } from "@/lib/convertSpeedToPace";
 import { useLocationStore } from "@/store/useLocationStore";
 import { useRunningStore } from "@/store/useRunningStore";
+import { Alert } from "react-native";
 
 const SoloRunning = () => {
   const { speed } = useSoloRunning();
@@ -12,7 +14,10 @@ const SoloRunning = () => {
   return (
     <>
       {myLocation && (
-        <RunningInfo seconds={seconds} rest={runDistance} speed={speed} />
+        <>
+          <RunningInfo seconds={seconds} rest={runDistance} speed={speed} />
+          {runDistance > 0 && <SoloRunningText />}
+        </>
       )}
     </>
   );
